+++ CHANGELOG.md
@@ -2,6 +2,8 @@ Unreleased
 ----------
 - Added support for test function attributes, e.g., `#[ignore]`
 - Added support for mutable `device` arguments to test functions
+- Removed restriction of expecting an argument specifically named
+  `device`
 - Introduced `nitrokey-test-state` crate containing state needed to
   achieve mutual exclusion of tests